itd: Add itd.h with basic format description
.itd is hrengin's new archive format.

It's purpose is to provide the most basic means of tightly packing bunch
of files together (slightly more than so).
It doesn't provide store paths, names, metadata, error
correction etc. Instead, those are supposed to be stored as files inside
archive or several archives (e.g. file 0 is directory tree, file 1 is
hash table, file 2 is tags, and archive 2 stores signatures of
each file in archive 1)
This way, .itd serves as a skeleton for hrengin packages.