doc: Reintroduce node-values
The idea is to combine the concepts of 'nodes' and 'values' into a
single one - there are only nodes which can have a single value and a
number of children.
In other words, what was considered a 'value' object before, now it
should become just a shorthand for a childless node.

i.e.

	[node child=value]

is same as

	[node [child=value]]

TODO: actual data structure for that
TODO: need a static_vector to reduce the size of such structure