graphics/gl-loader: fix extension loader
load_functions (now renamed to load_functions_ext) always returned
failure, changed it so it always returns success when there is nothing
to load.

It wasn't causing problems because it wasn't checked anywhere,
but it was in my stashed changes so I think it *did* cause problems
eventually (perhabs I tried to use some extensions?  don't remember
at this point).

Additionally did some cleanup, got rid of that PascalCase (snake_case is
the standard for awlib)