doc: rework node (get rid of values[])
Previously I commented on how I wanted to rework the graph so it contains
only nodes instead of separate nodes and values.
Now everything is a node (i.e. each node might have a value associated
with it and what was previously just a value now is a full-blown node)
It is a little bit (2x) less memory efficient, but is simpler and more
flexible.