rendertest: add aw::doc-based scene
It finally replaces the hardcoded texture lists and stuff. It is a draft
which needs to be refactored, but at least it allows me to replace the
old hard-to-read-and-modify txt format and opens the road for further
refactoring and experimentation.
It doesn't quite work (for some reason it adds only one object, thinks
that the others have an invalid material).

TODOs:
- split the resource system and the scene
- add a 'resource linker', so that load order doesn't matter (to allow
  resources to be put in multiple different files)
- use many different files instead of one big file
- decouple the render pipeline from the scene - rendering engine should
  be responsible only for rendering, it shouldn't impede its own use
  within a game engine; so it shouldn't provide some scene abstraction
  because *every* game engine already has its own abstraction