string: Add compose_v
My latest "improvement" broke awscript diagnostics, because it relied on
passing an array_view as a list of substitutions. Recent changes cause
it to be treated as a single argument, turning error diagnostics into
gibberish.

I'm adding compose_v (_v stands for "vector" or something like that) so
that I can have the best of both worlds: flexibility AND zero
allocations when using varargs (TBH, I use std::format now, so it's a
moot point)