utility: argv_parser: Add support for --option=x
It is something that wasn't supported before. It doesn't require many
changes. While some people might now want to support --option=x, I find
that it is more convenient to use propgrams that support both -
especially when you have to chain together programs that don't and the
ones that do.