lua: add registry
Main class for binding C++ classes to Lua. For now it only has a
push_reference methods, but I intend to add more stuff in the future.

Nothing is final yet, there are several design points I might change in
the future, but for now I use variadic templates to generate stuff in
compile-time to avoid lots of pointer chasing.