hudf: rename doc -> hudf
Not that long ago I renamed HDF (Hudd's Data Format) to aw::doc to avoid
confusion with other HDF (Hierarchical Data Format) which is used in
scientific community.

However, while browsing other repositories I realized that naming it
"doc" might also be confusing because the folder "<repo>/doc" is often
used to store various documentation. It also might be consufed for a
parser of the old Microsoft Word document format.

So I decided to rename it back to "Hudd's Data Format" but with a U in
it: HuDF.