meta/pp: fix preprocessor UB in tests
Usually I would've amended the previous commit, but it's useful to keep
it separate in this case as it makes easier to deal with this problem in
the future. Previously I didn't use a macro, instead I used
static_assert directly. However, using #if inside a function-like macro
invocation is UB for no good reason and Visual C++'s preprocessor
refuses to work correctly.

warning: C5101: use of preprocessor directive in function-like macro argument list is undefined behavior